backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
python-dotenv
aiogram
httpx
python-binance
redis
websockets==12.0
